+++ tests/test_plugins_cloudflare_ua_fix.py
@@ -1,0 +1,4 @@
+def test_plugin():
+    from reader._plugins.cloudflare_ua_fix import init
+
+    # TODO: actually test the plugin